Versão 1.442: Ajustes e correções nos formulários FPA, PIT e RIT: recurso autocompletar no campo de curso e disciplinas; Importação dos dados de FPA para PIT e de PIT para RIT; Melhorias no sincronizador Nambei em relação ao Consulta Roda: atualiza lista automaticamente ao clicar no botão; Funcionalidade de fixar ano/semestre; Todos os tipos de usuário com permissão de alterar ano/semestre por padrão (admin pode alterar);

+++ view/secretaria/cursos/disciplina.php
@@ -14,6 +14,17 @@ require PERMISSAO;
 require CONTROLLER . "/disciplina.class.php";
 $disciplina = new Disciplinas();
 
+if ($_GET["ajax"]) {
+    $arr = array();
+    header('Cache-Control: no-cache');
+    header('Content-type: application/xml; charset="utf-8"', true);
+
+    $dados = $disciplina->listToJSON(($_GET["codigo"]));
+
+    echo( json_encode($dados) );
+    die;
+}
+
 if ($_GET["opcao"] == 'delete') {
     $ret = $disciplina->delete($_GET["codigo"]);
     mensagem($ret['STATUS'], $ret['TIPO'], $ret['RESULTADO']);
